pyproject.toml
[tool.poetry]
name = "task_manager"
version = "0.1.0"
description = ""
authors = ["mxszinevich"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.2"
SQLAlchemy = "^1.4.42"
psycopg2 = "^2.9.5"
asyncpg = "^0.27.0"
alembic = "^1.8.1"
uvicorn = "^0.19.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
black = "^22.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]


[tool.isort]
profile = "black"
extend_skip = ["migrations", "venv"]
line_length = 120

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
    /(
        \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | venv
      | _build
      | buck-out
      | build
      | dist
    )/
'''

[tool.flake8]
max-line-length = 120
count = true
per-file-ignores = ['__init__.py:F401']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
